requirements.txt
huggingface_hub
transformers
torch
auto-gptq
autoawq
optimum
git+https://github.com/huggingface/peft.git